+++ tests/posit/enumeration.cpp
@@ -27,7 +27,7 @@ try {
 	const size_t nbits = 5;
 	const size_t es = 1;
 	// generate minpos
-	posit<nbits, es> p_minpos(uint64_t(0));
+	posit<nbits, es> p_minpos(0);
 	p_minpos++; // next posit to ZERO
 
 	// generate maxpos via INFINITE
@@ -42,7 +42,7 @@ try {
 	cout << "minpos : " << p_minpos << " maxpos : " << p_maxpos << " "  << endl;
 
 	cout << "State space enumeration" << endl;
-	posit<nbits, es> p(uint64_t(0));
+	posit<nbits, es> p(0);
 	for (int i = 0; i < uint32_t(1) << nbits; i++) {
 		cout << components_to_string(p) << endl;
 		p++;
